# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Nemo V2
  version: 1.0.0
...
paths:
...
  /api/v1/platform/role:
    post:
      tags:
        - FMS JS > Platform > Roles
      summary: Create Role
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                rolename: Customer Care Admin L3
                roletype: platform
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/platform/role/list:
    get:
      tags:
        - FMS JS > Platform > Roles
      summary: List roles
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/platform/role/{roleid}:
    get:
      tags:
        - FMS JS > Platform > Roles
      summary: Get Role Info
      parameters:
        - in: path
          name: roleid
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
    put:
      tags:
        - FMS JS > Platform > Roles
      summary: Update Role
      parameters:
        - in: path
          name: roleid
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                rolename: Customer Care Admin
                roletype: platform
                isenabled: true
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
    delete:
      tags:
        - FMS JS > Platform > Roles
      summary: Delete Role
      parameters:
        - in: path
          name: roleid
          required: true
          schema:
            type: string
            format: uuid
            description: Role ID to delete
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      roleid:
                        type: string
                        format: uuid
                      rolename:
                        type: string
                      deletedat:
                        type: string
                        format: date-time
                      deletedby:
                        type: string
                        format: uuid
                  msg:
                    type: string
                    example: Role deleted successfully
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: object
                    properties:
                      errcode:
                        type: string
                        enum:
                          [
                            "ROLE_IN_USE",
                            "ROLE_HAS_PERMISSIONS",
                            "ROLE_NOT_FOUND",
                            "CANNOT_DELETE_ADMIN_ROLE",
                          ]
                      errdata:
                        type: string
                  data:
                    type: "null"
...
  /api/v1/platform/user/{userid}/role:
    post:
      tags:
        - FMS JS > Platform > Users
      summary: Add User Platform Role
      parameters:
        - in: path
          name: userid
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                roleids:
                  - superadmin
                  - 9deea93f-3489-41fc-ab87-d55c777b0512
                roletype: platform
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/platform/user/{userid}/role/{roleid}:
    delete:
      tags:
        - FMS JS > Platform > Users
      summary: Delete User Role
      parameters:
        - in: path
          name: userid
          required: true
          schema:
            type: string
        - in: path
          name: roleid
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/fms/account/roles:
    get:
      tags:
        - FMS JS > FMS > AccountMangement > Role
      summary: Get Account Roles
      parameters:
        - in: path
          name: accountid
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/fms/account/role/{roleid}:
    get:
      tags:
        - FMS JS > FMS > AccountMangement > Role
      summary: Get Account Role Info
      parameters:
        - in: path
          name: accountid
          required: true
          schema:
            type: string
        - in: path
          name: roleid
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
    put:
      tags:
        - FMS JS > FMS > AccountMangement > Role
      summary: Edit Account Role
      parameters:
        - in: path
          name: accountid
          required: true
          schema:
            type: string
        - in: path
          name: roleid
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                rolename: testrole01
                roletype: account
                isenabled: true
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/fms/account/role:
    post:
      tags:
        - FMS JS > FMS > AccountMangement > Role
      summary: Create Account Role
      parameters:
        - in: path
          name: accountid
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                rolename: testrole01
                roletype: account
                isenabled: true
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...